extracted string-formatting and string-concatination related functions out of string functions class.
