+++ src/ExAs/Assertions/EnumerableAssertion.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using ExAs.Results;
 using ExAs.Utils;
+using static ExAs.Utils.StringFormattingFunctions;
 
 namespace ExAs.Assertions
 {
@@ -16,7 +17,7 @@ namespace ExAs.Assertions
         public Result AssertEnumerable(IEnumerable<TElement> enumerable)
         {
             var valueAssertionResult = itemAssertion.AssertValue(enumerable);
-            string log = StringFunctions.HangingIndent("Enumerable<".Add(typeof(TElement).Name).Add(">: "), valueAssertionResult.actualValueString);
+            string log = HangingIndent("Enumerable<".Add(typeof(TElement).Name).Add(">: "), valueAssertionResult.actualValueString);
             return new Result(valueAssertionResult.succeeded, log, valueAssertionResult.expectationString);
         }
     }
